Move DefaultRayRuntimeFactory into runtime module.
