+++ java/api/src/main/java/org/ray/api/Ray.java
@@ -2,7 +2,6 @@ package org.ray.api;
 
 import java.util.List;
 import org.ray.api.id.UniqueId;
-import org.ray.api.runtime.DefaultRayRuntimeFactory;
 import org.ray.api.runtime.RayRuntime;
 import org.ray.api.runtime.RayRuntimeFactory;
 
@@ -17,7 +16,15 @@ public final class Ray extends RayCall {
    * Initialize Ray runtime with the default runtime implementation.
    */
   public static void init() {
-    init(new DefaultRayRuntimeFactory());
+    try {
+      Class clz = Class.forName("org.ray.runtime.DefaultRayRuntimeFactory");
+      RayRuntimeFactory factory = (RayRuntimeFactory) clz.newInstance();
+      init(factory);
+
+    } catch (Exception e) {
+      throw new RuntimeException("Failded to initialize Ray API.");
+    }
+
   }
 
   /**
