+++ java/runtime/src/main/java/org/ray/runtime/RayNativeRuntime.java
@@ -17,7 +17,7 @@ import org.ray.runtime.gcs.StateStoreProxyImpl;
 import org.ray.runtime.raylet.RayletClient;
 import org.ray.runtime.raylet.RayletClientImpl;
 import org.ray.runtime.runner.RunManager;
-import org.ray.api.util.RayLog;
+import org.ray.runtime.util.RayLog;
 
 /**
  * native runtime for local box and cluster run.
