+++ java/runtime/src/main/java/org/ray/runtime/WorkerContext.java
@@ -2,7 +2,7 @@ package org.ray.runtime;
 
 import org.ray.api.WorkerMode;
 import org.ray.api.id.UniqueId;
-import org.ray.api.config.RayConfig;
+import org.ray.runtime.util.RayConfig;
 import org.ray.runtime.task.TaskSpec;
 
 public class WorkerContext {
