+++ java/runtime/src/main/java/org/ray/runtime/functionmanager/LocalFunctionManager.java
@@ -6,7 +6,7 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.ray.api.id.UniqueId;
 import org.ray.runtime.task.FunctionArg;
 import org.ray.runtime.util.Serializer;
-import org.ray.api.util.RayLog;
+import org.ray.runtime.util.RayLog;
 
 /**
  * local function manager which pulls remote functions on demand.
