+++ java/runtime/src/main/java/org/ray/runtime/functionmanager/NativeRemoteFunctionManager.java
@@ -10,8 +10,8 @@ import org.ray.runtime.gcs.KeyValueStoreLink;
 import org.ray.runtime.util.FileUtil;
 import org.ray.runtime.util.JarLoader;
 import org.ray.runtime.util.Sha1Digestor;
-import org.ray.api.util.SystemUtil;
-import org.ray.api.util.RayLog;
+import org.ray.runtime.util.SystemUtil;
+import org.ray.runtime.util.RayLog;
 
 /**
  * native implementation of remote function manager.
