+++ java/runtime/src/main/java/org/ray/runtime/raylet/RayletClientImpl.java
@@ -16,7 +16,7 @@ import org.ray.runtime.generated.TaskLanguage;
 import org.ray.runtime.task.FunctionArg;
 import org.ray.runtime.task.TaskSpec;
 import org.ray.runtime.util.UniqueIdHelper;
-import org.ray.api.util.RayLog;
+import org.ray.runtime.util.RayLog;
 
 public class RayletClientImpl implements RayletClient {
