+++ java/runtime/src/main/java/org/ray/runtime/util/MethodId.java
@@ -14,9 +14,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.WeakHashMap;
 import org.objectweb.asm.Type;
-import org.ray.api.util.StringUtil;
-import org.ray.api.util.RayLog;
-
 
 /**
  * An instance of RayFunc is a lambda.
