+++ java/runtime/src/main/java/org/ray/runtime/util/NetworkUtil.java
@@ -1,4 +1,4 @@
-package org.ray.api.util;
+package org.ray.runtime.util;
 
 import java.io.IOException;
 import java.net.DatagramSocket;
