+++ java/runtime/src/main/java/org/ray/runtime/util/Sha1Digestor.java
@@ -3,9 +3,6 @@ package org.ray.runtime.util;
 import java.nio.ByteBuffer;
 import java.security.MessageDigest;
 
-import org.ray.api.util.StringUtil;
-import org.ray.api.util.RayLog;
-
 public class Sha1Digestor {
 
   private static final ThreadLocal<MessageDigest> md = ThreadLocal.withInitial(() -> {
