+++ java/runtime/src/main/java/org/ray/runtime/util/StringUtil.java
@@ -1,4 +1,4 @@
-package org.ray.api.util;
+package org.ray.runtime.util;
 
 import java.nio.charset.Charset;
 import java.util.ArrayList;
