+++ java/test/src/main/java/org/ray/api/benchmark/RayBenchmarkTest.java
@@ -11,7 +11,7 @@ import org.ray.api.Ray;
 import org.ray.api.RayActor;
 import org.ray.api.RayObject;
 import org.ray.api.annotation.RayRemote;
-import org.ray.api.util.RayLog;
+import org.ray.runtime.util.RayLog;
 
 public abstract class RayBenchmarkTest<T> implements Serializable {
