+++ java/test/src/main/java/org/ray/api/test/LambdaUtilsTest.java
@@ -12,7 +12,7 @@ import org.ray.api.function.RayFunc0;
 import org.ray.api.function.RayFunc1;
 import org.ray.api.function.RayFunc3;
 import org.ray.runtime.util.MethodId;
-import org.ray.api.util.RayLog;
+import org.ray.runtime.util.RayLog;
 
 public class LambdaUtilsTest {
